Trriger処理ボタン完成 hpとスケールの関係を作る

+++ Game/Collision/CollisionManager.cpp
@@ -80,7 +80,7 @@ void CollisionManager::CheckCollisionPair(Collider* colliderA, Collider* collide
 	}*/
 	// 判定対象AとBの座標
 	Vector3 posA, posB;
-	int radiusA, radiusB;
+	float radiusA, radiusB;
 
 	// colliderAの座標
 	posA = colliderA->GetWorldPosition();
@@ -94,7 +94,7 @@ void CollisionManager::CheckCollisionPair(Collider* colliderA, Collider* collide
 
 	float p2b = (posB.x - posA.x) * (posB.x - posA.x) + (posB.y - posA.y) * (posB.y - posA.y) +
 		(posB.z - posA.z) * (posB.z - posA.z);
-	int r2r = (radiusA + radiusB) * (radiusA + radiusB);
+	float r2r = (radiusA + radiusB) * (radiusA + radiusB);
 	/*  if (((colliderA->GetCollisonAttribute() & colliderB->GetCollisionMask())!=0) ||
 		  ((colliderB->GetCollisonAttribute() & colliderA->GetCollisionMask()))!=0) {
 	  return;
